+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.8.0] - ALPHA
 
-> **Note**<br>This release is currently in alpha test phase and therefore not yet released for production use.
+> **Note**<br>This release is currently in alpha test phase and therefore not yet recommended for production use.
 
 ### Update Notes
 
@@ -20,6 +20,8 @@ Please update all contacts after completing the installation. You can do that si
 ### Added
 
 - Standing revocations now show reason to managers (#11)
+- Standing revocations now shows related contact labels
+- Shows contact labels for all effective standing requests
 - Contact sets are now visible on the admin site
 - New button for updating all contacts from the admin site
 - Faction info for characters and corporations on standing pages